+++ src/events/recieve/reciveRmq.ts
@@ -1,7 +1,7 @@
 import amqp, { Channel, Connection } from 'amqplib'
 
 import { Env } from '../../config/env'
-import winstonLogger from '../../lib/WinstonLogger'
+import winstonLogger from '../../lib/winstonLogger'
 import { RmqError } from '../../shared/errors/rmqError'
 import { ISendController } from '../../shared/interfaces/rmq/sendRmqController'
 import { deserializeMessage } from '../shared/serializeMessage'
